requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
